refactor: Extract metadata schemas to subpackage and add PaginationQuery
