+++ backend/src/expenses_counter/modules/routers/schemas/base/metadata/__init__.py
@@ -1,0 +1,6 @@
+"""Metadata schemas module."""
+
+__all__ = ("PaginationMetadata", "PaginationQuery")
+
+from .pagination_metadata import PaginationMetadata
+from .pagination_query import PaginationQuery
