+++ backend/src/expenses_counter/modules/routers/schemas/base/metadata/pagination_query.py
@@ -1,0 +1,15 @@
+"""Query model for pagination."""
+
+__all__ = ("PaginationQuery",)
+
+from pydantic import Field
+
+from ..query import BaseQueryModel
+
+
+class PaginationQuery(BaseQueryModel):
+    """Query parameters for pagination."""
+
+    limit: int = Field(default=100, ge=1, le=100, description="The number of items to return")
+    offset: int = Field(default=0, ge=0, description="The number of items to skip")
+    sort_order: str = Field(default="asc", description="The order to sort the items by")
